pyproject.toml
[project]
name = "jobauto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.26",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "python-dotenv>=1.1.1",
    "youtube-search>=2.1.2",
    "youtube-search-python>=1.6.6",
]
